+++ lib/honeybadger/configuration.rb
@@ -6,7 +6,7 @@ module Honeybadger
                :params_filters, :project_root, :port, :protocol, :proxy_host, :proxy_pass,
                :proxy_port, :proxy_user, :secure, :use_system_ssl_cert_chain, :framework,
                :user_information, :rescue_rake_exceptions, :source_extract_radius,
-               :send_request_session, :debug].freeze
+               :send_request_session, :debug, :fingerprint].freeze
 
     # The API key for your project, found on the project edit form.
     attr_accessor :api_key
@@ -100,6 +100,9 @@ module Honeybadger
 
     # A Proc object used to send notices asynchronously
     attr_writer :async
+
+    # A Proc object used to generate optional fingerprint
+    attr_writer :fingerprint
 
     DEFAULT_PARAMS_FILTERS = %w(password password_confirmation).freeze
 
@@ -261,6 +264,24 @@ module Honeybadger
       @async
     end
     alias :async? :async
+
+    # Public: Generate custom fingerprint (optional)
+    #
+    # block - An optional block returning calculated fingerprint
+    #
+    # Examples
+    #
+    #   config.fingerprint = Proc.new { |notice| ... }
+    #
+    #   config.fingerprint do |notice|
+    #     Digest::SHA1.hexdigest([notice[:error_class], notice[:component], notice[:backtrace].to_s].join(':'))
+    #   end
+    #
+    # Returns configured fingerprint generator (should respond to #call(notice))
+    def fingerprint
+      @fingerprint = Proc.new if block_given?
+      @fingerprint
+    end
 
     def port
       @port || default_port